chore: review and store another htpass in the source code ; sure
We need to connect the htpasswd generation to the .env* files.
This webpage could have no password at all and it would be OK,
so I'm not too concerned about the HORRIBLE SECURITY ISSUE
of having a .htpasswd in the source tree.

Anyway, let's remove it eventually, but don't worry.